feat(types): centralize consultation type definitions
